distutils.dir_util.copy_tree
Python's shutil.copytree is weirdly limited. It won't let you copy to a
destination that already exists. We weren't using it right, but the
tests didn't show it because the lib dir didn't have any subdirs. Use
the copy_tree function in distutils instead, and add a subdir in our
tests.