default rule is now just "rule"
We were using a field called "default_rule" to specify which rule
counted as the default for a module. That was pretty ugly, and it
bloated the hello-world example.

One option was to allow modules to include rule fields directly, the
idea being that these fields applied to the default rule of that module.
That was concise, but the problem is that we also want some module
fields (e.g. "imports") to apply to *all* rules within the module. So it
would be ambiguous which fields affected only the default rule.

The current compromise is to specify the default rule with just "rule:"
(instead of "rule foo:"). Fields within that rule clearly apply only to
that rule, the hello-world example is still reasonably concise, and we
avoid needing to bless a particular rule name as "the one that means
default".

    module hello_world_example:
        type: git
        url: http://example.com
        rev: master
        rule:
            build: make
            export: out/