allow imports to environment variables
Normally imports are given as paths where we should dump rule outputs.
Interpret a path beginning with '$' as an environment variable import.
The rule is still built (or fetched from cache), but the outputs are put
in a temp directory, and the given environment variable is set to the
path to that directory before any build commands are run.