reimplement export_tree with git-checkout
export_tree() now takes an optional previous_hash (defaults to a hash
representing the empty commit). The export is then implemented as a
checkout from the previous_hash to the hash being exported. That means
we get status checking for free. We will also use this feature to keep
imports clean in local modules.