start fixing test.sh tests
We're holding off on reimplementing environment variable imports and
default rules, since it's less clear how those should work now, so
remove those features from the tests. Previous import cleanup is still
unimplemented, so the second call to peru in the tests is failing.

The first test fails about half the time on the submodule check, so I
need to track that down.