fix the submodules test failure
What was happening was that the path module was pointing to a git repo.
Unfortunately, our git cache implementation has magic behavior when we
try to check in other git repos. (It ignores their .git directories and
submodules.) If the path module was imported first, that magic state
polluted the git module, because we weren't cleaning the index properly.
Add a line to import_tree() to fix this problem.

The "half the time" part resulted from the fact that we depended on
dictionary order to apply imports, which was random. Sort the imports to
avoid this in the future.

TODO: Figure out how to handle caching other git repos. This affets both
path modules and any rules whose build command includes a git clone. *shudder*