sidestep the import cleanup with git-checkout --force
Import cleanup is tricky. We can keep a mapping of all the trees we've
previously imported in the local module. But there's no good way to
associate present imports with past ones. (Names can change, and more
than one target can be imported to the same path.) There's also the
problem of running into false-positive conflicts when an imported file
changes ownership from one target to another, and the new owner gets
imported before the old one is cleaned.

What we actually need is a way of aggregating all of our imports into a
single tree, so that git can checkout the whole thing all at once. In
the meantime, git-checkout --force ignores the cleanup issue entirely
and gets us to a state where all the tests in test.sh are passing.

The git rewrite is finally ready to push to master.