stop using flags in the plugin interface
When invoking a plugin, pass all fields first as plain alternative keys
and values. Terminate this list with a "--" argument. Then pass the
command name and positional parameters. For example:

  git_plugin.py url $url rev $rev -- fetch $dest $cache_path

This makes sense as we move towards adding a "reup" command. All plugin
invocations will pass in the entire set of fields, so we should put
those first. After all of those are consumed, the plugin can figure out
what to do with the command.