start edit_yaml.py
We want to be able to edit YAML files in place for the "reup" feature.
To do that, the first step is to get the start and end indices of a
given field. In this diff, we define an initial function that
recursively walks the list of YAML parse events and builds lists and
dictionaries out of them as appropriate, so we can find the tokens we're
looking for.