always fetch on a reup
In theory we could avoid the fetch if your reup target is a tag or a raw
hash, but there's not really any use case for such a target. Also, tags
can actually change.

TODO: The git_plugin code is getting really nasty. Clean it up as part
of writing a shared plugin main function.