add a failing test for edit_yaml
There's a bug where, if the last field in a module is a dict (like rule
or imports), the end_mark of that field will actually be the beginning
of the next toplevel section. We insert new fields on the line after,
which means we're inserting them in the wrong section.