create Runtime, major refactor
Summary:
The `Runtime` class now contains most of the state that peru needs to pass
around, including module/rule definitions, config flags, and the cache
instance. In several cases, functions that took all these parameters separately
now just take one `Runtime`. As a result of this change, the `Resolver` no
longer needs to hold any state of its own, and all its methods have been turned
into functions.

The original motiviation for all these changes is that we wanted to make it
easier for the user to keep the entire cache dir (not just the plugins cache
dir) somewhere in his home folder. For that, we needed to pull the `tmp_dir`
functionality out of `Cache`, because peru shouldn't be creating tons of temp
directories in your home folder if your project lives somewhere else. We needed
a new class to represent the state stored in the .peru dir itself (like overrides),
which usually lives in the project even if the cache doesn't. Now `Runtime` is
this class. The `PERU_PLUGINS_CACHE` env var is gone, and only
`PERU_CACHE` remains.

Test Plan: Modified a couple integration tests.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D14