factor the out KeyVal class out of Cache
Summary:
Our storage for overrides is very similar to the keval storage we do in Cache,
but so far we've been using a different mechanism for it. In preparation for
unifying them, separate the KeyVal class out of Cache and beef it up with a few
more functions.

Test Plan: Separate and expand the existing KeyVal unit test.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D18