delete override.py
Summary:
That file was a lot of code doing very little. Replace it with the newly
exposed KeyVal class. KeyVal allows more granular access to the values stored
on disk, and it avoids keeping duplicated copies of data in memory that could
get out of sync. (Though of course, we could do that in the future for
performance if we need.) Also, it has a nice dictionary-like interface.

Test Plan: Existing override tests cover this.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D19