make the override command handle relative paths correctly
Summary:
When adding overrides on the command line, users can supply relative paths
based on their current working dir. However, when users later run a build that
reads these override paths, there's no guarantee that their working dir will be
the same. We could always translate to absolute paths for storage, but that
would make it impossible to do convenient things like copy an entire workspace
of projects to a different place on disk (their absolute override paths would
no longer be valid).

To keep the advantages of relative paths, reinterpret them relative to the
project root when they're stored. And when overrides are listed out for the user,
they should again be reinterpreted relative to the users current directory.

Test Plan: A big new integration test for this.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D20