always set plugins' working dir to the project root
Summary:
Plugins can receive relative paths in their fields (for example, the path to a
git repo). These paths come from the peru.yaml file, and they should always be
interpreted relative to the directory that contains that file. To make sure
that happens, explicitly set the working directory of plugins to the project
root whenever we shell out.

Test Plan: Update one of the integration tests to exercise this.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D21