add the 'files' field to rules and modules
Summary:
For a while now we've been using `build` rules to handle file filtering, by
copying files into an out directory and then exporting it. That's pretty hacky,
and it's worse when it needs to be portable. The new `files:` field replaces
these hacks with native globbing from `pathlib`. These globs are resolved
relative to the `export` path.

Also update our peru.yaml file to take advantage of the new field.

Test Plan: New unit test for `Cache` and integration tests for `Rule`.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D30