fix tests that were cluttering the repo with temp files
Summary:
We were using relative paths in a few tests that made files get created in the
peru repo itself. Gross. Clean these up. Add a check in `test.sh` so that we
will notice if this happens again.

Test Plan: Run the new `test.sh` with the old master. Observe failure.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D34