make sure export_tree() recreates deleted files
Summary:
In D39, we relaxed `export_tree` to allow deleted files, but we forgot
to handle the case where those files are both deleted and unmodified between
trees. In that case, `read-tree -mu` leaves them missing, instead of recreating
them. Add an extra `checkout-index --all` to make sure these files get
recreated.

Test Plan:
Add a new test for this behavior, and confirm that it was failing
before this diff. Also add a contents check where one was missing.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D51