make validate_third_party.sh compatible with OSX
Summary:
The `--tmp-dir` flag isn't supported. Stop using `/tmp/perutest`
entirely, because we want to create that with broad write permissions in
`tests/shared.py`, and it's best not to duplicate that logic.

Test Plan:
Run it on a Mac. It would be great to have CI that supports OSX. I
think it's possible to hack that in with Travis, but we'd need to do some
whacky stuff with branches?

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D52