add the --parallel option to limit simultaneous fetches
Summary:
Set the default limit to 10. Use an `asyncio` semaphore to limit the
parallelism. Docopt is wonderfully flexible with flag arguments, so the user
can write any one of `-p N`, `-pN`, `--parallel N`, or `--parallel=N`.

Test Plan:
Brief testing on my apartment wifi suggests that 10 is just as fast
as 15 for small GitHub repos. A Google search confirmed that 10 is pretty.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D59