support lists of imports as well as maps
Summary:
The easiest way to represent imports is as a map, but sometimes you
want to import the same module more than once. (We particularly want this in
testing.) This diff adds support for lists of imports as well as maps. Those
look like this:

```
imports:
    - foo: path/
    - bar: path/
```

We also create the `Imports` data structure so that code outside the parser
doesn't need to worry about this distinction.

Closes https://github.com/buildinspace/peru/issues/38.

Test Plan: New unit tests.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D60