get rid of the toplevel 'plugins:' field
Summary:
This feature didn't turn out the way we liked. It would be neat if
modules could fetch their own plugin definitions, but that would require some
kind of early fetch, so we went with a static version that requires checking in
plugins. But even that was going to cause a lot of complexity when we enable
recursive projects, which would involve unpacking the calling tree every time a
submodule was fetched.

Instead, we're going to fall back on the simpler design, where plugins have to
be installed on the user's system before they're used, just like peru itself.
Eventually we'll add plenty more levers for this, like default plugin dirs for
the system and for the user, and some env vars to extend that.

Test Plan:
Tests related to the plugins path have been deleted. We'll need to rewrite them
when we have those levers. Arguably we should have some more tests that
directly exercise the lower level methods of plugins.py.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D63