check the version of python in bin/peru
Summary:
This happens before we import the peru libs and hit a bunch of sytax
errors on older Pythons. We'll need to keep `bin/peru` backwards compatible,
but that's fine since there should never be any significant logic in there.

Test Plan:
Try to run peru from a python3.2 and python2.7. Get a clean error in
both cases.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D65