make sure DEBUG_PARALLEL_COUNT is decremented on a plugin error
Summary:
This global is used in tests, to keep track of how many jobs run at
once. Jobs are supposed to increment it when they start and decrement it
when they finish. But if the plugin returned an error code, we were
throwing and failing to decrement that code.

This didn't repro for me locally, because the only tests that checked
the counter were happening early in my test run, but it caused Travis
failures because of a quirk in test run order.

This diff fixes the issue by putting the decrement in a finally clause,
and it extends the plugin tests so that all of them sanity check this
counter.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D130