sync with -q in validate_third_party.sh
Summary:
Travis likes to pretend that its tests are running in a terminal, but it
doesn't support the ANSI control codes that our fancy display uses, so
we get a spew of crap in the build logs when peru runs inside of
validate_third_party.sh. Use `-q` to avoid the problem.

Reviewers: sean

Differential Revision: https://phabricator.buildinspace.com/D127