stop LocalModule from creating a .peru dir by default
Summary:
Instead, only create the dir when we need to actually write a
`lastimports` file. This is a short term fix for our parser tests
inadvertently creating a `.peru` directory in our tests dir. With this
fix, D145 should be able to remove `.peru` from our `.gitignore` file.

In the long term, it's clear that we have an organization problem. The
parser should obviously not be calling code that writes to disk. I think
the best solution will be to tear up the LocalModule class. It's a
holdover from when we thought we would be doing much more complicated
things to override dirs, like applying imports and running build
commands. Going forward, override dirs are going to be a much simpler
concept, and we can probably separate the parser from the code that
deals with the local project dir in the process of simplifying it.

Tracking issue: https://github.com/buildinspace/peru/issues/72

Test Plan:
`arc patch D145`, remove `.peru` from `.gitignore`, clean the repo, and
run tests. Tests pass without any warning about dirty files.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D146