get rid of the `build` field
Summary:
With recursive modules coming up, the `build` field is starting to look
like a security vulnerability. It's also a bit of a portability issue --
projects that use it will usually only work on the one platform where
their build commands are valid (see
https://github.com/buildinspace/peru/issues/17). It's looking like the
Right Thing is to get rid of it entirely and leave all the building to
whatever tool is calling peru. This diff removes all the `build` code
and adds a specific error message for projects that use it.

I realized in the course of implementing this that there was a bug in
resolver.py, where we delegated building to `LocalModule` instead of
using `Rule` to do it. (This was because `export` directories needed to
be handled specially as the cwd of subsequent commands.) But
`LocalModule` didn't handle the `files` field, so that was probably
broken with respect to overrides. This diff significantly simplifies
LocalModule, using it only to get an input file tree, so we no longer
treat it differently in resolver.py (making progress on
https://github.com/buildinspace/peru/issues/72).

Test Plan:
Adding a test for the deprecation message, and removing `build` from
several tests.

Reviewers: sean

Reviewed By: sean

Differential Revision: https://phabricator.buildinspace.com/D158